Made player-position stuff doubles. Accurate movement now!

+++ src/main/java/mobagame/core/game/InGamePlayer.java
@@ -8,8 +8,8 @@ import mobagame.launcher.GameScreen;
 
 public class InGamePlayer {
 	private int playerID;
-	private int x;
-	private int y;
+	private double x;
+	private double y;
 	private Character character;
 	private int phyPow;
 	private int abiPow;
@@ -196,11 +196,11 @@ public class InGamePlayer {
 		return currentMana;
 	}
 
-	public int getX() {
+	public double getX() {
 		return x;
 	}
 
-	public int getY() {
+	public double getY() {
 		return y;
 	}
 
@@ -234,11 +234,11 @@ public class InGamePlayer {
 		this.playerShape = playerShape;
 	}
 
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
 	public void levelUp() {
